Move layout layer and append options
