+++ layout/write.go
@@ -17,6 +17,25 @@ import (
 	"github.com/google/go-containerregistry/pkg/v1/layout"
 )
 
+type AppendOption func(*appendOptions)
+
+type appendOptions struct {
+	withoutLayers bool
+	annotations   map[string]string
+}
+
+func WithoutLayers() AppendOption {
+	return func(i *appendOptions) {
+		i.withoutLayers = true
+	}
+}
+
+func WithAnnotations(annotations map[string]string) AppendOption {
+	return func(i *appendOptions) {
+		i.annotations = annotations
+	}
+}
+
 // AppendImage mimics GGCR's AppendImage in that it appends an image to a `layout.Path`,
 // but the image appended does not include any layers in the `blobs` directory.
 // The returned image will return layers when Layers(), LayerByDiffID(), or LayerByDigest() are called,
